% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targ_dbtbl.R
\name{db_targ}
\alias{db_targ}
\title{Create a new db_targ object}
\usage{
db_targ(
  con,
  schema,
  table_name,
  query = "select * from {`schema`}.{`table_name`}"
)
}
\arguments{
\item{con}{A database connection}

\item{schema}{Name of schema containing the table}

\item{table_name}{Name of the table}

\item{query}{Query to execute against the table. Defaults to selecting all rows and columns}
}
\value{
A \code{db_targ} object
}
\description{
Create a new db_targ object
}
