% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_db_targ.R
\name{parse_db_targ}
\alias{parse_db_targ}
\title{Partse a \code{db_targ} stored as a .rds file}
\usage{
parse_db_targ(file_path)
}
\arguments{
\item{file_path}{Path to a .rds file}
}
\value{
A \code{db_targ} object
}
\description{
Partse a \code{db_targ} stored as a .rds file
}
