% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_db_targ.R
\name{read_db_targ}
\alias{read_db_targ}
\title{Read a \code{db_targ} table}
\usage{
read_db_targ(con, targ)
}
\arguments{
\item{con}{A database connection}

\item{targ}{A \code{db_targ} object}
}
\value{
A data.frame
}
\description{
Read a \code{db_targ} table
}
