% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_db_targ.R
\name{save_db_targ}
\alias{save_db_targ}
\title{Save a \code{db_targ} object. Convenience wrapper for base \code{saveRDS}}
\usage{
save_db_targ(targ, file_path)
}
\arguments{
\item{targ}{A \code{db_targ} object}

\item{file_path}{File to save the \code{db_targ} object to. Must be a .rds file}
}
\value{
file_path, invisibly
}
\description{
Save a \code{db_targ} object. Convenience wrapper for base \code{saveRDS}
}
