% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db_targ.R
\name{write_db_targ}
\alias{write_db_targ}
\title{Write a data frame to a database using a \code{db_targ}}
\usage{
write_db_targ(con, df, targ, ...)
}
\arguments{
\item{con}{A database connection}

\item{df}{A data frame}

\item{targ}{A \code{db_targ} object}

\item{...}{Arbitrary arguments to pass to \code{DBI::dbWriteTable}}
}
\description{
Write a data frame to a database using a \code{db_targ}
}
